% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{build.test.case}
\alias{build.test.case}
\title{Comparison of 3 strategies}
\usage{
build.test.case(nportfolios, nassets, start_date, end_date,
  riskFreeRate = 0, collapse = c("daily", "weekly", "monthly",
  "quarterly", "annual"), api_key)
}
\arguments{
\item{nportfolios}{An integer, number of portfolios to be generated}

\item{nassets}{An integer, number of stock in a portfolio}

\item{start_date}{A Date object, format="YYYY-MM-DD". Retrieve data rows on and after the specified start date.}

\item{end_date}{A Date object, format="YYYY-MM-DD". Retrieve data rows up to and including the specified end date.}

\item{collapse}{Change the sampling frequency of the returned data. Default is none; i.e., data is returned in its original granularity.}

\item{api_key}{A string, authentication toket for a Quandl user.}
}
\value{
list, training=market training dataset, test=market test dataset, sharp=performance of sharp ratio on test dataset, rational=performance of rational investor on test dataset, random=performance of random investor on test dataset.
}
\description{
A function is used to compare 3 different stretagies: sharp ration, rational invetor and random investor based on training dataset.
Then, it applies derived weights to test dataset to see the performance of those strategies.
}
