% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketDataAPI.R
\name{get.marketData.random}
\alias{get.marketData.random}
\title{Random end of day US Stock prices}
\usage{
get.marketData.random(nassets, start_date, end_date,
  collapse = c("daily", "weekly", "monthly", "quarterly", "annual"),
  api_key, excludedCodes = NULL, skipNA = TRUE, recursive = FALSE)
}
\arguments{
\item{nassets}{An integer, number of stock in a portfolio}

\item{start_date}{A Date object, format="YYYY-MM-DD". Retrieve data rows on and after the specified start date.}

\item{end_date}{A Date object, format="YYYY-MM-DD". Retrieve data rows up to and including the specified end date.}

\item{collapse}{Change the sampling frequency of the returned data. Default is none; i.e., data is returned in its original granularity.}

\item{api_key}{A string, authentication toket for a Quandl user.}

\item{excludedCodes}{A string or list of strings, stock codes that should be excluded from random sampling.}

\item{skipNA}{Boolean, if skipNA=TRUE, an assert that has NA in its date range is excluded.}

\item{recursive}{Boolean, if recursive=TRUE, a function will execute until exact predefined number of stocks is found.}
}
\value{
list(data=dataframe of stocks in portfolio, missingCode=vector of codes that do not satisfy input params)
}
\description{
Retrieves N number of random stocks from the Quandl Dataset.
}
\note{
If recursive=TRUE & skipNA=TRUE, a function will search for N asserts that satisfy start_date, end_date and collapse arguments. Therefore, eventually, it may go out of requests allowed by Quandle.
}
\references{
Also see get.marketData function for additional information.
}
