% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketdataAPI.R
\name{get.marketData}
\alias{get.marketData}
\title{End of day US Stock Proces}
\usage{
get.marketData(code, start_date, end_date, collapse = c("daily",
  "weekly", "monthly", "quarterly", "annual"), api_key)
}
\arguments{
\item{code}{string or list of strings: Code a.k.a ticker symbol on Quandle.}

\item{start_date}{A Date object, format="YYYY-MM-DD". Retrieve data rows on and after the specified start date.}

\item{end_date}{A Date object, format="YYYY-MM-DD". Retrieve data rows up to and including the specified end date.}

\item{collapse}{Change the sampling frequency of the returned data. Default is none; i.e., data is returned in its original granularity.}

\item{api_key}{Authentication toket for a Quandl user.}
}
\value{
dataframe of returns per stock
}
\description{
Retrieves Data from the Quandl Dataset endpoint and formats
}
\note{
Due to Quandle database specification, all ".", " " symbols are represented as "_".
}
\examples{
\dontrun{
code <- c("AAPL", "FB", "GOOG", "C", "A", "INTC", "T", "WMT", "TXN")
start_date <- "2017-12-31"
end_date <- "2018-12-31"
collapse <- "daily"
api_key <- "User123"s
market_date <- get.marketDate (code=code, start_date=start_date, end_date=end_date, collapse=collapse, api_key=api_key)
}
}
\references{
This R package uses the Quandl API. For more information go to \url{https://www.quandl.com/docs/api}. For more help on the package itself go to \url{https://www.quandl.com/help/r}.
}
