% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketDataAPI.R
\name{get.portfolios.random}
\alias{get.portfolios.random}
\title{Random set of portfolios of end of day US Stock prices}
\usage{
get.portfolios.random(nportfolios, nassets, ...)
}
\arguments{
\item{nportfolios}{An integer, number of portfolios to be generated}

\item{...}{See get.marketData.random}
}
\description{
Generate N numberb of randomly sampled, unique portfolios
}
