% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkowitzUW.R
\docType{package}
\name{MarkowitzUW}
\alias{MarkowitzUW}
\alias{MarkowitzUW-package}
\title{MarkowitzUW}
\description{
A package designed for the Advnced R course at University of Warsaw, summer semester 2018-2019
}
\details{
The MarkowitzUW provides two categories of functions: Market data API and strategies.
}
\section{Market data API}{

The Market data API functions are used to retrieves financial data from the Quandl Dataset. It's based on End of day US Stock Prices.
}

\section{Strategies}{

There are tree strategies used to evaluate the performance of portfolio: sharp ration, rational invester and random inverster stretagies.
}

