% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{plot.strategies}
\alias{plot.strategies}
\title{Plot the output of 3 strategies}
\usage{
\method{plot}{strategies}(df)
}
\arguments{
\item{df}{build.test.case output}
}
\value{
list of function to plot.
}
\description{
A function is used to plot 3 different stretagies: sharp ration, rational invetor and random investor based on training dataset.
}
\note{
In order to plot a portfolio performance, you need to select a particular portfolio from a list and run it.
}
