% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{squared}
\alias{squared}
\title{Square a number}
\usage{
squared(x)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
The square of the input
}
\description{
Take in any numeric value and square it
}
