% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{strategy.random}
\alias{strategy.random}
\title{Random investor strategy}
\usage{
strategy.random(marketData)
}
\arguments{
\item{marketData}{data.frame of list of data.frame, marketData is used to calculate weights}
}
\value{
vector of weights of each stock in a portfolio
}
\description{
The random investor strategy is based on random investor behaviour. All weights are assigned randomly
}
