% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{strategy.rational}
\alias{strategy.rational}
\title{Rational investor strategy}
\usage{
strategy.rational(marketData)
}
\arguments{
\item{marketData}{data.frame of list of data.frame, marketData is used to calculate weights}
}
\value{
vector of weights of each stock in a portfolio
}
\description{
The rational investor strategy is based on the following model: a stock with highest average rate of returns receives 50% wegith, second highest is set to 25% weight, the rest receives equal share of remaning 25%. Note, than onle possitive rate of returns is concerned.
}
